+++ tests/gold_tests/remap/all_acl_combinations.py
@@ -17,126 +17,6 @@ Verify remap.config acl behavior.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-import os
-import io
-import re
-import pathlib
-import inspect
-import tempfile
-from yaml import load, dump
-from yaml import CLoader as Loader
-from typing import List, Tuple
-from deactivate_ip_allow import all_deactivate_ip_allow_tests
-
-Test.Summary = '''
-Verify remap.config acl behavior.
-'''
-
-
-class Test_remap_acl:
-    """Configure a test to verify remap.config acl behavior."""
-
-    _ts_counter: int = 0
-    _server_counter: int = 0
-    _client_counter: int = 0
-
-    def __init__(
-            self, name: str, replay_file: str, ip_allow_content: str, deactivate_ip_allow: bool, acl_matching_policy: int,
-            acl_configuration: str, named_acls: List[Tuple[str, str]], expected_responses: List[int]):
-        """Initialize the test.
-
-        :param name: The name of the test.
-        :param replay_file: The replay file to be used.
-        :param ip_allow_content: The ip_allow configuration to be used.
-        :param deactivate_ip_allow: Whether to deactivate the ip_allow filter.
-        :param acl_configuration: The ACL configuration to be used.
-        :param named_acls: The set of named ACLs to configure and use.
-        :param expect_responses: The in-order expected responses from the proxy.
-        """
-        self._replay_file = replay_file
-        self._ip_allow_content = ip_allow_content
-        self._deactivate_ip_allow = deactivate_ip_allow
-        self._acl_matching_policy = acl_matching_policy
-        self._acl_configuration = acl_configuration
-        self._named_acls = named_acls
-        self._expected_responses = expected_responses
-
-        tr = Test.AddTestRun(name)
-        self._configure_server(tr)
-        self._configure_traffic_server(tr)
-        self._configure_client(tr)
-
-    def _configure_server(self, tr: 'TestRun') -> None:
-        """Configure the server.
-
-        :param tr: The TestRun object to associate the server process with.
-        """
-        name = f"server-{Test_remap_acl._server_counter}"
-        server = tr.AddVerifierServerProcess(name, self._replay_file)
-        Test_remap_acl._server_counter += 1
-        self._server = server
-
-    def _configure_traffic_server(self, tr: 'TestRun') -> None:
-        """Configure Traffic Server.
-
-        :param tr: The TestRun object to associate the Traffic Server process with.
-        """
-
-        name = f"ts-{Test_remap_acl._ts_counter}"
-        ts = tr.MakeATSProcess(name, enable_cache=False, enable_tls=True)
-        Test_remap_acl._ts_counter += 1
-        self._ts = ts
-
-        ts.Disk.records_config.update(
-            {
-                'proxy.config.diags.debug.enabled': 1,
-                'proxy.config.diags.debug.tags': 'http|url|remap|ip_allow',
-                'proxy.config.http.push_method_enabled': 1,
-                'proxy.config.http.connect_ports': self._server.Variables.http_port,
-                'proxy.config.url_remap.acl_matching_policy': self._acl_matching_policy,
-            })
-
-        remap_config_lines = []
-        if self._deactivate_ip_allow:
-            remap_config_lines.append('.deactivatefilter ip_allow')
-
-        # First, define the name ACLs (filters).
-        for name, definition in self._named_acls:
-            remap_config_lines.append(f'.definefilter {name} {definition}')
-        # Now activate them.
-        for name, _ in self._named_acls:
-            remap_config_lines.append(f'.activatefilter {name}')
-
-        remap_config_lines.append(f'map / http://127.0.0.1:{self._server.Variables.http_port}/ {self._acl_configuration}')
-        ts.Disk.remap_config.AddLines(remap_config_lines)
-        ts.Disk.ip_allow_yaml.AddLines(self._ip_allow_content.split("\n"))
-
-    def _configure_client(self, tr: 'TestRun') -> None:
-        """Run the test.
-
-        :param tr: The TestRun object to associate the client process with.
-        """
-
-        name = f"client-{Test_remap_acl._client_counter}"
-        p = tr.AddVerifierClientProcess(name, self._replay_file, http_ports=[self._ts.Variables.port])
-        Test_remap_acl._client_counter += 1
-        p.StartBefore(self._server)
-        p.StartBefore(self._ts)
-
-        if self._expected_responses == [None, None]:
-            # If there are no expected responses, expect the Warning about the rejected ip.
-            self._ts.Disk.diags_log.Content += Testers.ContainsExpression(
-                "client '127.0.0.1' prohibited by ip-allow policy", "Verify the client rejection warning message.")
-
-            # Also, the client will complain about the broken connections.
-            p.ReturnCode = 1
-
-        else:
-            codes = [str(code) for code in self._expected_responses]
-            p.Streams.stdout += Testers.ContainsExpression(
-                '.*'.join(codes), "Verifying the expected order of responses", reflags=re.DOTALL | re.MULTILINE)
-
-
 ALLOW_GET_AND_POST = f'''
 ip_allow:
   - apply: in
@@ -193,25 +73,6 @@ ip_allow:
 replay_proxy_response writes the given replay file (which expects a single GET & POST client-request)
 with the given proxy_response value. This is only used to support the below table test.
 '''
-
-
-def replay_proxy_response(filename, replay_file, get_proxy_response, post_proxy_response):
-    current_dir = os.path.dirname(inspect.getfile(inspect.currentframe()))
-    path = os.path.join(current_dir, filename)
-    data = None
-    with open(path) as f:
-        data = load(f, Loader=Loader)
-        for session in data["sessions"]:
-            for transaction in session["transactions"]:
-                method = transaction["client-request"]["method"]
-                if method == "GET":
-                    transaction["proxy-response"]["status"] = 403 if get_proxy_response == None else get_proxy_response
-                elif method == "POST":
-                    transaction["proxy-response"]["status"] = 403 if post_proxy_response == None else post_proxy_response
-                else:
-                    raise Exception("Expected to find GET or POST request, found %s", method)
-    with open(replay_file, "w") as f:
-        f.write(dump(data))
 
 
 # yapf: disable
@@ -309,52 +170,3 @@ all_acl_combinations = [
 # yapf: enable
 
 all_acl_combination_tests = [dict(zip(keys, test)) for test in all_acl_combinations]
-"""
-Test all acl combinations
-"""
-for idx, test in enumerate(all_acl_combination_tests):
-    (_, replay_file_name) = tempfile.mkstemp(suffix="acl_table_test_{}.replay".format(idx))
-    replay_proxy_response(
-        "base.replay.yaml",
-        replay_file_name,
-        test["GET response"],
-        test["POST response"],
-    )
-    Test.Summary = "table test {0}".format(idx)
-    Test_remap_acl(
-        "{0} {1} {2}".format(test["inline"], test["named_acl"], test["ip_allow"]),
-        replay_file=replay_file_name,
-        ip_allow_content=test["ip_allow"],
-        deactivate_ip_allow=False,
-        acl_matching_policy=0 if test["policy"] == "explicit" else 1,
-        acl_configuration=test["inline"],
-        named_acls=[("acl", test["named_acl"])] if test["named_acl"] != "" else [],
-        expected_responses=[test["GET response"], test["POST response"]],
-    )
-
-"""
-Test all ACL combinations
-"""
-for idx, test in enumerate(all_deactivate_ip_allow_tests):
-    try:
-        test["deactivate_ip_allow"]
-    except:
-        print(test)
-    (_, replay_file_name) = tempfile.mkstemp(suffix="deactivate_ip_allow_table_test_{}.replay".format(idx))
-    replay_proxy_response(
-        "base.replay.yaml",
-        replay_file_name,
-        test["GET response"],
-        test["POST response"],
-    )
-    Test.Summary = "table test {0}".format(idx)
-    Test_remap_acl(
-        "{0} {1} {2}".format(test["inline"], test["named_acl"], test["ip_allow"]),
-        replay_file=replay_file_name,
-        ip_allow_content=test["ip_allow"],
-        deactivate_ip_allow=test["deactivate_ip_allow"],
-        acl_matching_policy=0 if test["policy"] == "explicit" else 1,
-        acl_configuration=test["inline"],
-        named_acls=[("acl", test["named_acl"])] if test["named_acl"] != "" else [],
-        expected_responses=[test["GET response"], test["POST response"]]
-    )
