+++ tests/gold_tests/remap/remap_acl.test.py
@@ -18,7 +18,13 @@ Verify remap.config acl behavior.
 #  limitations under the License.
 
 import os
+import io
 import re
+import pathlib
+import inspect
+import tempfile
+from yaml import load, dump
+from yaml import CLoader as Loader
 from typing import List, Tuple
 
 Test.Summary = '''
@@ -154,7 +160,6 @@ test_ip_allow_optional_methods = Test_remap_acl(
     named_acls=[],
     expected_responses=[200, 200, 403, 403, 403])
 
-"""
 test_ip_allow_optional_methods = Test_remap_acl(
     "Verify if no ACLs match, ip_allow.yaml is used.",
     replay_file='remap_acl_get_allowed.replay.yaml',
@@ -328,4 +333,76 @@ test_named_acl_deny = Test_remap_acl(
     acl_configuration='',
     named_acls=[('deny', '@action=deny @method=HEAD @method=POST')],
     expected_responses=[200, 403, 403, 403])
+
+def replay_proxy_response(filename, replay_file, get_proxy_response, post_proxy_response):
+    current_dir = os.path.dirname(inspect.getfile(inspect.currentframe()))
+    path = os.path.join(current_dir, filename)
+    data = None
+    with open(path) as f:
+        data = load(f, Loader=Loader)
+        for session in data["sessions"]:
+            for transaction in session["transactions"]:
+                method = transaction["client-request"]["method"]
+                if method == "GET":
+                    transaction["proxy-response"]["status"] = 403 if get_proxy_response == None else get_proxy_response
+                elif method == "POST":
+                    transaction["proxy-response"]["status"] = 403 if post_proxy_response == None else post_proxy_response
+                else:
+                    raise Exception("Expected to find GET or POST request, found %s", method)
+    with open(replay_file, "w") as f:
+        f.write(dump(data))
+
+
+
+from deactivate_ip_allow import all_deactivate_ip_allow_tests
+from all_acl_combinations import all_acl_combination_tests
+
 """
+Test all acl combinations
+"""
+for idx, test in enumerate(all_acl_combination_tests):
+    (_, replay_file_name) = tempfile.mkstemp(suffix="acl_table_test_{}.replay".format(idx))
+    replay_proxy_response(
+        "base.replay.yaml",
+        replay_file_name,
+        test["GET response"],
+        test["POST response"],
+    )
+    Test.Summary = "table test {0}".format(idx)
+    Test_remap_acl(
+        "{0} {1} {2}".format(test["inline"], test["named_acl"], test["ip_allow"]),
+        replay_file=replay_file_name,
+        ip_allow_content=test["ip_allow"],
+        deactivate_ip_allow=False,
+        acl_matching_policy=0 if test["policy"] == "explicit" else 1,
+        acl_configuration=test["inline"],
+        named_acls=[("acl", test["named_acl"])] if test["named_acl"] != "" else [],
+        expected_responses=[test["GET response"], test["POST response"]],
+    )
+
+"""
+Test all ACL combinations
+"""
+for idx, test in enumerate(all_deactivate_ip_allow_tests):
+    try:
+        test["deactivate_ip_allow"]
+    except:
+        print(test)
+    (_, replay_file_name) = tempfile.mkstemp(suffix="deactivate_ip_allow_table_test_{}.replay".format(idx))
+    replay_proxy_response(
+        "base.replay.yaml",
+        replay_file_name,
+        test["GET response"],
+        test["POST response"],
+    )
+    Test.Summary = "table test {0}".format(idx)
+    Test_remap_acl(
+        "{0} {1} {2}".format(test["inline"], test["named_acl"], test["ip_allow"]),
+        replay_file=replay_file_name,
+        ip_allow_content=test["ip_allow"],
+        deactivate_ip_allow=test["deactivate_ip_allow"],
+        acl_matching_policy=0 if test["policy"] == "explicit" else 1,
+        acl_configuration=test["inline"],
+        named_acls=[("acl", test["named_acl"])] if test["named_acl"] != "" else [],
+        expected_responses=[test["GET response"], test["POST response"]]
+    )
